Linenoise - begin clean-up, split up files and convert to classes
