+++ tools/shell/linenoise/CMakeLists.txt
@@ -1,0 +1,13 @@
+include_directories(include)
+
+add_library(
+        duckdb_linenoise
+        OBJECT
+        highlighting.cpp
+        history.cpp
+        linenoise.cpp
+        linenoise-c.cpp
+        terminal.cpp)
+set(SHELL_SOURCES
+        ${SHELL_SOURCES} $<TARGET_OBJECTS:duckdb_linenoise>
+        PARENT_SCOPE)
