+++ tools/shell/linenoise/highlighting.cpp
@@ -1,0 +1,3 @@
+#include "linenoise.hpp"
+#include "linenoise.h"
+
